Enum roead::aamp::Parameter

pub enum Parameter {

    Bool(bool),
    F32(f32),
    I32(i32),
    Vec2(Vector2f),
    Vec3(Vector3f),
    Vec4(Vector4f),
    Color(Color),
    String32(FixedSafeString<32>),
    String64(Box<FixedSafeString<64>>),
    Curve1(Box<[Curve; 1]>),
    Curve2(Box<[Curve; 2]>),
    Curve3(Box<[Curve; 3]>),
    Curve4(Box<[Curve; 4]>),
    BufferInt(Vec<i32>),
    BufferF32(Vec<f32>),
    String256(Box<FixedSafeString<256>>),
    Quat(Quat),
    U32(u32),
    BufferU32(Vec<u32>),
    BufferBinary(Vec<u8>),
    StringRef(String),

}

Parameter.

Note that unlike agl::utl::Parameter the name is not stored as part of the parameter class in order to make the parameter logic simpler and more efficient.

Variants

Bool(bool)

Boolean.

F32(f32)

Float.

I32(i32)

Int.

Vec2(Vector2f)

2D vector.

Vec3(Vector3f)

3D vector.

Vec4(Vector4f)

4D vector.

Color(Color)

Color.

String32(FixedSafeString<32>)

String (max length 32 bytes).

String64(Box<FixedSafeString<64>>)

String (max length 64 bytes).

Curve1(Box<[Curve; 1]>)

A single curve.

Curve2(Box<[Curve; 2]>)

Two curves.

Curve3(Box<[Curve; 3]>)

Three curves.

Curve4(Box<[Curve; 4]>)

Four curves.

BufferInt(Vec<i32>)

Buffer of signed ints.

BufferF32(Vec<f32>)

Buffer of floats.

String256(Box<FixedSafeString<256>>)

String (max length 256 bytes).

Quat(Quat)

Quaternion.

U32(u32)

Unsigned int.

BufferU32(Vec<u32>)

Buffer of unsigned ints.

BufferBinary(Vec<u8>)

Buffer of binary data.

StringRef(String)

String (no length limit).

Implementations

impl Parameter

pub fn as_bool(&self) -> Result<bool>

Get the inner bool value.

pub fn as_mut_bool(&mut self) -> Result<&mut bool>

Get a mutable reference to the inner bool.

pub fn into_bool(self) -> Result<bool>

Extract the inner bool value.

pub fn as_f32(&self) -> Result<f32>

Get the inner f32 value.

pub fn as_mut_f32(&mut self) -> Result<&mut f32>

Get a mutable reference to the inner f32.

pub fn into_f32(self) -> Result<f32>

Extract the inner f32 value.

pub fn as_i32(&self) -> Result<i32>

Get the inner i32 value.

pub fn as_mut_i32(&mut self) -> Result<&mut i32>

Get a mutable reference to the inner i32.

pub fn as_int<T>(&self) -> Result<T>
where T: Copy + 'static, i32: AsPrimitive<T>, u32: AsPrimitive<T>,

Get the inner value as an integer of any type. Casts the value using as where necessary. Note that this is subject to all the normal risks of casting with as.

pub fn as_num<T>(&self) -> Result<T>
where T: Copy + 'static, i32: AsPrimitive<T>, u32: AsPrimitive<T>, f32: AsPrimitive<T>,

Get the inner value as a number of any type. Casts the value using as where necessary. Note that this is subject to all the normal risks of casting with as.

pub fn into_i32(self) -> Result<i32>

Extract the inner i32 value.

pub fn as_vec2(&self) -> Result<&Vector2f>

Get the inner Vector2f value.

pub fn as_mut_vec2(&mut self) -> Result<&mut Vector2f>

Get a mutable reference to the inner Vector2f.

pub fn into_vec2(self) -> Result<Vector2f>

Extract the inner Vector2f value.

pub fn as_vec3(&self) -> Result<&Vector3f>

Get the inner Vector3f value.

pub fn as_mut_vec3(&mut self) -> Result<&mut Vector3f>

Get a mutable reference to the inner Vector3f.

pub fn into_vec3(self) -> Result<Vector3f>

Extract the inner Vector3f value.

pub fn as_vec4(&self) -> Result<&Vector4f>

Get the inner Vector4f value.

pub fn as_mut_vec4(&mut self) -> Result<&mut Vector4f>

Get a mutable reference to the inner Vector4f.

pub fn into_vec4(self) -> Result<Vector4f>

Extract the inner Vector4f value.

pub fn as_buffer_f32(&self) -> Result<&[f32]>

Get the inner BufferF32 value.

pub fn as_mut_buffer_f32(&mut self) -> Result<&mut Vec<f32>>

Get a mutable reference to the inner BufferF32.

pub fn into_buffer_f32(self) -> Result<Vec<f32>>

Extract the inner BufferF32 value.

pub fn as_buffer_int(&self) -> Result<&[i32]>

Get the inner BufferI32 value.

pub fn as_mut_buffer_int(&mut self) -> Result<&mut Vec<i32>>

Get a mutable reference to the inner BufferI32.

pub fn into_buffer_int(self) -> Result<Vec<i32>>

Extract the inner BufferI32 value.

pub fn as_color(&self) -> Result<&Color>

Get the inner color value.

pub fn as_mut_color(&mut self) -> Result<&mut Color>

Get a mutable reference to the inner color.

pub fn into_color(self) -> Result<Color>

Extract the inner color value.

pub fn as_string32(&self) -> Result<&FixedSafeString<32>>

Get the inner String32 value.

pub fn as_mut_string32(&mut self) -> Result<&mut FixedSafeString<32>>

Get a mutable reference to the inner String32.

pub fn into_string32(self) -> Result<FixedSafeString<32>>

Extract the inner String32 value.

pub fn as_string64(&self) -> Result<&FixedSafeString<64>>

Get the inner String64 value.

pub fn as_mut_string64(&mut self) -> Result<&mut FixedSafeString<64>>

Get a mutable reference to the inner String64.

pub fn into_string64(self) -> Result<FixedSafeString<64>>

Extract the inner String64 value.

pub fn as_string256(&self) -> Result<&FixedSafeString<256>>

Get the inner String256 value.

pub fn as_mut_string256(&mut self) -> Result<&mut FixedSafeString<256>>

Get a mutable reference to the inner String256.

pub fn into_string256(self) -> Result<FixedSafeString<256>>

Extract the inner String256 value.

pub fn as_curve1(&self) -> Result<&[Curve; 1]>

Get the inner Curve1 value.

pub fn as_mut_curve1(&mut self) -> Result<&mut [Curve; 1]>

Get a mutable reference to the inner Curve1.

pub fn into_curve1(self) -> Result<[Curve; 1]>

Extract the inner Curve1 value.

pub fn as_curve2(&self) -> Result<&[Curve; 2]>

Get the inner Curve2 value.

pub fn as_mut_curve2(&mut self) -> Result<&mut [Curve; 2]>

Get a mutable reference to the inner Curve2.

pub fn into_curve2(self) -> Result<[Curve; 2]>

Extract the inner Curve2 value.

pub fn as_curve3(&self) -> Result<&[Curve; 3]>

Get the inner Curve3 value.

pub fn as_mut_curve3(&mut self) -> Result<&mut [Curve; 3]>

Get a mutable reference to the inner Curve3.

pub fn into_curve3(self) -> Result<[Curve; 3]>

Extract the inner Curve3 value.

pub fn as_curve4(&self) -> Result<&[Curve; 4]>

Get the inner Curve4 value.

pub fn as_mut_curve4(&mut self) -> Result<&mut [Curve; 4]>

Get a mutable reference to the inner Curve4.

pub fn into_curve4(self) -> Result<[Curve; 4]>

Extract the inner Curve4 value.

pub fn as_quat(&self) -> Result<&Quat>

Get the inner Quat value.

pub fn as_mut_quat(&mut self) -> Result<&mut Quat>

Get a mutable reference to the inner Quat.

pub fn into_quat(self) -> Result<Quat>

Extract the inner Quat value.

pub fn as_u32(&self) -> Result<u32>

Get the inner u32 value.

pub fn as_mut_u32(&mut self) -> Result<&mut u32>

Get a mutable reference to the inner u32.

pub fn into_u32(self) -> Result<u32>

Extract the inner u32 value.

pub fn as_buffer_u32(&self) -> Result<&[u32]>

Get the inner u32 buffer value.

pub fn as_mut_buffer_u32(&mut self) -> Result<&mut [u32]>

Get a mutable reference to the inner u32 buffer.

pub fn into_buffer_u32(self) -> Result<Vec<u32>>

Extract the inner u32 buffer value.

pub fn as_buffer_binary(&self) -> Result<&[u8]>

Get the inner binary buffer value.

pub fn as_mut_buffer_binary(&mut self) -> Result<&mut [u8]>

Get a mutable reference to the inner binary buffer.

pub fn into_buffer_binary(self) -> Result<Vec<u8>>

Extract the inner binary buffer value.

pub fn as_string_ref(&self) -> Result<&str>

Get the inner string value.

pub fn as_mut_string_ref(&mut self) -> Result<&mut str>

Get a mutable reference to the inner string.

pub fn into_string_ref(self) -> Result<String>

Extract the inner string value.

pub fn as_safe_string<const N: usize>(&self) -> Result<FixedSafeString<N>>

Get the value from any string type. May truncate.

impl Parameter

pub fn as_str(&self) -> Result<&str>

Returns a string slice if the parameter is any string type.