Struct roead::aamp::NameTable

pub struct NameTable<'a> { /* private fields */ }

Available on crate features aamp-names and aamp only.

A table of names that is used to recover original names in binary parameter archives which store only name hashes.

Because binary parameter archives only store CRC32 hashes of structure names, recovering the original names – which is useful for converting archives to a human-readable format – requires the use of a name table.

When serializing to YAML, by default roead will use a table that contains strings from Breath of the Wild’s executable.

Implementations

impl<'a> NameTable<'a>

pub fn new(botw_strings: bool) -> NameTable<'a>

Create a new name table, optionally including default BOTW strings.

pub fn add_name(&self, name: impl Into<Cow<'a, str>>)

Add a known string to the name table.

pub fn add_name_with_hash(&self, name: impl Into<Cow<'a, str>>, hash: u32)

Add a known string to the name table if you already know the hash (to avoid computing it).

pub fn add_name_str<'s: 'a>(&'a self, name: &'s str)

Add a known string to the name table.

pub fn get_name( &self, hash: u32, index: usize, parent_hash: u32, ) -> Option<&Cow<'_, str>>

Tries to guess the name that is associated with the given hash and index (of the parameter / object / list in its parent).

The table is automatically updated with any newly found names if an indice-based guess was necessary.