Struct roead::aamp::ParameterIO

pub struct ParameterIO {
    pub version: u32,
    pub data_type: String,
    pub param_root: ParameterList,
}

Available on crate feature aamp only.

Parameter IO. This is the root parameter list and the only structure that can be serialized to or deserialized from a binary parameter archive.

Fields
version: u32

Data version (not the AAMP format version). Typically 0.

data_type: String

Data type identifier. Typically โxmlโ.

param_root: ParameterList

Root parameter list.

Implementations

impl ParameterIO

pub fn read<R: Read + Seek>(reader: R) -> Result<ParameterIO>

Read a parameter archive from a binary reader.

pub fn from_binary(data: impl AsRef<[u8]>) -> Result<ParameterIO>

Load a parameter archive from binary data.

Note: If and only if the yaz0 feature is enabled, this function automatically decompresses the data when necessary.

impl ParameterIO

pub fn from_text(text: impl AsRef<str>) -> Result<Self>

Available on crate feature yaml only.

Parse ParameterIO from YAML text.

pub fn to_text(&self) -> String

Available on crate feature yaml only.

Serialize the parameter IO to YAML.

impl ParameterIO

pub fn write<W: Write + Seek>(&self, writer: W) -> Result<()>

Serialize the parameter IO to binary using the given writer.

pub fn to_binary(&self) -> Vec<u8>

Serialize the parameter IO to in-memory bytes.

impl ParameterIO

pub fn new() -> Self

Create a new empty parameter IO.

pub fn with_data_type(self, data_type: impl Into<String>) -> ParameterIO

Builder-like method to set the data type.

pub fn with_version(self, version: u32) -> ParameterIO

Builder-like method to set the data version.

pub fn with_object<N: Into<Name>>( self, name: N, object: ParameterObject, ) -> ParameterIO

Builder-like method to add a new parameter object.

pub fn with_objects<N: Into<Name>, I: IntoIterator<Item = (N, ParameterObject)>>( self, iter: I, ) -> ParameterIO

Builder-like method to add multiple parameter objects from an iterator.

pub fn with_list<N: Into<Name>>( self, name: N, list: ParameterList, ) -> ParameterIO

Builder-like method to add a new parameter list.

pub fn with_lists<N: Into<Name>, I: IntoIterator<Item = (N, ParameterList)>>( self, iter: I, ) -> ParameterIO

Builder-like method to add multiple parameter lists from an iterator.

pub fn with_root(self, list: ParameterList) -> ParameterIO

Builder-like method to set the root parameter list.