Trait roead::aamp::ParameterListing

pub trait ParameterListing {
    // Required methods
    fn lists(&self) -> &ParameterListMap;
    fn lists_mut(&mut self) -> &mut ParameterListMap;
    fn objects(&self) -> &ParameterObjectMap;
    fn objects_mut(&mut self) -> &mut ParameterObjectMap;

    // Provided methods
    fn list<N: Into<Name>>(&self, name: N) -> Option<&ParameterList> { ... }
    fn list_mut<N: Into<Name>>(&mut self, name: N) -> Option<&mut ParameterList> { ... }
    fn set_list<N: Into<Name>>(&mut self, name: N, list: ParameterList) { ... }
    fn object<N: Into<Name>>(&self, name: N) -> Option<&ParameterObject> { ... }
    fn object_mut<N: Into<Name>>(
        &mut self,
        name: N
    ) -> Option<&mut ParameterObject> { ... }
    fn set_object<N: Into<Name>>(&mut self, name: N, object: ParameterObject) { ... }
}

Trait abstracting over ParameterList and ParameterIO. Useful since a parameter IO is all but interchangeable with the root list.

Required Methods

fn lists(&self) -> &ParameterListMap

Returns a map of parameter lists.

fn lists_mut(&mut self) -> &mut ParameterListMap

Returns a mutable map of parameter lists.

fn objects(&self) -> &ParameterObjectMap

Returns a map of parameter objects.

fn objects_mut(&mut self) -> &mut ParameterObjectMap

Returns a mutable map of parameter objects.

Provided Methods

fn list<N: Into<Name>>(&self, name: N) -> Option<&ParameterList>

Get a parameter list by name or hash.

fn list_mut<N: Into<Name>>(&mut self, name: N) -> Option<&mut ParameterList>

Get a mutable reference to a parameter list by name or hash.

fn set_list<N: Into<Name>>(&mut self, name: N, list: ParameterList)

Set a parameter list by name or hash.

fn object<N: Into<Name>>(&self, name: N) -> Option<&ParameterObject>

Get a parameter object by name or hash.

fn object_mut<N: Into<Name>>(&mut self, name: N) -> Option<&mut ParameterObject>

Get a mutable reference to a parameter object by name or hash.

fn set_object<N: Into<Name>>(&mut self, name: N, object: ParameterObject)

Set a parameter object by name or hash.