Module roead::yaz0

Available on crate feature yaz0 only.

Bindings for the oead::yaz0 module, which supports Yaz0 decompression and fast compression (using syaz0).

Structs

CompressOptions

Yaz0 compression options.

Header

The header of Yaz0 compressed data.

Functions

compress

Compress data with default compression settings (no alignment, compression level 7).

compress_if

Compress data conditionally, if an associated path has a Yaz0-associated file extension (starts with s, but does not equal sarc). Returns a Cow which contains the original data if the data does not need to be compressed, or containing the compressed data otherwise.

compress_with_options

Compress data with custom compression settings.

decompress

Decompress Yaz0 data to vector.

decompress_if

Conditionally decompress Yaz0 data to a vector. Returns a Cow which contains the original data if the data is not Yaz0 compressed or decompression fails, or containing the decompressed data otherwise.

decompress_into

Decompress Yaz0 data into an existing buffer, returning the number of bytes written.

decompress_unchecked⚠

Decompress Yaz0 data into an existing buffer, returning the number of bytes written.

get_header

Get the header of Yaz0 compressed data, if it exists.